+++ lib/capsium/thor_ext.rb
@@ -60,7 +60,7 @@ module Capsium
         return if error.is_a?(Errno::EPIPE)
         raise if ENV["VERBOSE"] || !config.fetch(:exit_on_failure, true)
 
-        message = error.message.to_s
+        message = +error.message.to_s
         message.prepend("[#{error.class}] ") if message.empty? || !error.is_a?(Thor::Error)
 
         config[:shell]&.say_error(message, :red)